+++ examples/volumetric/legosurface.py
@@ -2,17 +2,14 @@
 Colors correspond to the volume's scalar.
 Try also:
 > vedo --lego data/embryo.tif"""
-# https://matplotlib.org/users/colormaps.html
 from vedo import *
 
-vol = Volume(dataurl+'embryo.tif').printHistogram(logscale=True)
+vol = Volume(dataurl+'embryo.tif')
 
-vol.crop(back=0.5) # crop 50% from neg. y
+vol.crop(back=0.50) # crop 50% from neg. y
 
 # show lego blocks whose value is between vmin and vmax
-lego = vol.legosurface(vmin=60, cmap='seismic')
-
-# make colormap start at 40
-lego.addScalarBar(horizontal=True, c='k')
+lego = vol.legosurface(vmin=20, vmax=None, boundary=False)
+lego.cmap('seismic', on='cells', vmin=0, vmax=127).addScalarBar()
 
 show(lego, __doc__, axes=1, viewup='z').close()